[BROOKLYN-161] Move osgi test bundles to a separate project
