+++ utils/test-bundles/entities/src/main/java/brooklyn/osgi/tests/SimpleLocation.java
