+++ utils/test-bundles/entities/src/main/resources/brooklyn/osgi/tests/icon.gif
