+++ utils/test-bundles/more-entities-v1/src/main/java/brooklyn/osgi/tests/more/MoreLocation.java
